Add gateway project
