+++ build.gradle
@@ -131,15 +131,11 @@ project(':yggdrash-node') {
 
     dependencies {
         compile project(":yggdrash-core")
+        compile("org.springframework.boot:spring-boot-starter-web")
         compile("org.springframework.boot:spring-boot-starter-undertow")
         compile("org.springframework.boot:spring-boot-starter-actuator")
-        compile("org.springframework.boot:spring-boot-starter-websocket")
         compile("org.springframework.cloud:spring-cloud-starter:2.0.0.RELEASE")
         compile("com.github.briandilley.jsonrpc4j:jsonrpc4j:1.5.3")
-        compile("org.webjars:sockjs-client:1.1.2")
-        compile("org.webjars:stomp-websocket:2.3.3")
-        compile("org.webjars:bootstrap:3.3.7")
-        compile("org.webjars:webjars-locator-core")
         annotationProcessor("org.springframework.boot:spring-boot-configuration-processor")
         runtime('org.springframework.boot:spring-boot-devtools')
         testCompile('org.springframework.boot:spring-boot-starter-test')
@@ -149,6 +145,7 @@ project(':yggdrash-node') {
 
     configurations {
         compile.exclude module: 'spring-boot-starter-tomcat'
+        compile.exclude module: 'spring-security-rsa'
     }
 
     docker {
@@ -157,6 +154,10 @@ project(':yggdrash-node') {
         name "yggdrash/${bootJar.baseName}"
         files bootJar.archivePath
         buildArgs(['JAR_FILE': "${bootJar.archiveName}"])
+    }
+
+    jar {
+        enabled = true
     }
 
     bootJar {
@@ -173,6 +174,31 @@ project(':yggdrash-node') {
                 }
             }
         }
+    }
+}
+
+project(':yggdrash-gateway') {
+    apply plugin: 'io.spring.dependency-management'
+    apply plugin: 'org.springframework.boot'
+
+    dependencies {
+        compile project(":yggdrash-node")
+        compile("org.springframework.boot:spring-boot-starter-undertow")
+        compile("org.springframework.boot:spring-boot-starter-websocket")
+        compile("org.webjars:sockjs-client:1.1.2")
+        compile("org.webjars:stomp-websocket:2.3.3")
+        compile("org.webjars:bootstrap:3.3.7")
+        compile("org.webjars:webjars-locator-core")
+        testCompile('org.springframework.boot:spring-boot-starter-test')
+        testCompile project(":yggdrash-core").sourceSets.test.output
+    }
+
+    configurations {
+        compile.exclude module: 'spring-boot-starter-tomcat'
+    }
+
+    bootJar {
+        launchScript()
     }
 }
