+++ yggdrash-gateway/src/main/java/io/yggdrash/gateway/controller/ExceptionControllerAdvice.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.yggdrash.node.controller;
+package io.yggdrash.gateway.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
