+++ yggdrash-gateway/src/main/resources/static/app.js
