+++ yggdrash-gateway/src/main/resources/static/favicon.ico
