+++ yggdrash-gateway/src/main/resources/static/index.html
