+++ yggdrash-gateway/src/main/resources/static/main.css
