+++ yggdrash-node/src/main/java/io/yggdrash/node/NodeScheduler.java
@@ -114,6 +114,9 @@ class NodeScheduler {
     }
 
     private boolean isSeedPeer(List<String> seedPeerList) {
+        if (seedPeerList == null) {
+            return true;
+        }
         String nodeUriWithoutPubKey = nodeManager.getNodeUri();
         nodeUriWithoutPubKey = nodeUriWithoutPubKey.substring(nodeUriWithoutPubKey.indexOf("@"));
         for (String seedPeer : seedPeerList) {
