--- yggdrash-node/src/main/java/io/yggdrash/node/api/AdminDto.java
@@ -1,34 +1,0 @@
-package io.yggdrash.node.api;
-
-public class AdminDto {
-
-    private String header;
-
-    private String signature;
-
-    private String body;
-
-    public String getHeader() {
-        return header;
-    }
-
-    public void setHeader(String header) {
-        this.header = header;
-    }
-
-    public String getSignature() {
-        return signature;
-    }
-
-    public void setSignature(String signature) {
-        this.signature = signature;
-    }
-
-    public String getBody() {
-        return body;
-    }
-
-    public void setBody(String body) {
-        this.body = body;
-    }
-}
