+++ yggdrash-node/src/main/java/io/yggdrash/node/api/BlockApi.java
@@ -6,7 +6,7 @@ import com.googlecode.jsonrpc4j.JsonRpcParam;
 import com.googlecode.jsonrpc4j.JsonRpcService;
 import io.yggdrash.core.exception.InternalErrorException;
 import io.yggdrash.core.exception.NonExistObjectException;
-import io.yggdrash.node.controller.BlockDto;
+import io.yggdrash.node.api.dto.BlockDto;
 
 @JsonRpcService("/api/block")
 public interface BlockApi {
