+++ yggdrash-node/src/main/java/io/yggdrash/node/api/BlockApiImpl.java
@@ -6,7 +6,7 @@ import io.yggdrash.core.blockchain.BranchGroup;
 import io.yggdrash.core.blockchain.BranchId;
 import io.yggdrash.core.exception.InternalErrorException;
 import io.yggdrash.core.exception.NonExistObjectException;
-import io.yggdrash.node.controller.BlockDto;
+import io.yggdrash.node.api.dto.BlockDto;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
