+++ yggdrash-node/src/main/java/io/yggdrash/node/api/dto/AccountDto.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2018 Akashic Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.yggdrash.node.api.dto;
+
+import io.yggdrash.core.wallet.Account;
+import org.apache.commons.codec.binary.Hex;
+
+public class AccountDto {
+    private String address;
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public static AccountDto createBy(Account account) {
+        AccountDto dto = new AccountDto();
+        dto.setAddress(Hex.encodeHexString(account.getAddress()));
+        return dto;
+    }
+}
