+++ yggdrash-node/src/test/java/io/yggdrash/node/api/TransactionApiImplTest.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2018 Akashic Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.yggdrash.node.api;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -6,7 +22,7 @@ import io.yggdrash.core.blockchain.BranchId;
 import io.yggdrash.core.blockchain.TransactionHusk;
 import io.yggdrash.core.contract.ContractTx;
 import io.yggdrash.core.wallet.Wallet;
-import io.yggdrash.node.controller.TransactionDto;
+import io.yggdrash.node.api.dto.TransactionDto;
 import org.junit.Before;
 import org.junit.Test;
 import org.slf4j.Logger;
