+++ yggdrash-node/src/test/resources/branch/ygg-id.json
@@ -7,5 +7,5 @@
   "genesis": {},
   "timestamp": "00000166c837f0c9",
   "owner": "c91e9d46dd4b7584f0b6348ee18277c10fd7cb94",
-  "signature": "1b571242de13285d93efc90807e9f51de3ee5a39c987efcc4d166f4f54fe4f980866acfa92f8b4d9d0eb0015a7c5a1c9256bf645d9bcdd2dd91bf7087e95517614"
+  "signature": "1cbb2f53e0c6fdc5c250e96dddef6197ee49ba7a410c3583e14971a325aa59ac951f19672e19970b155a15f7266b129f7be319319e37b3b2d704a6f9a8dbdf2a63"
 }
